Fixed several warnings. Add projects revisions. Fixed delete non exclusive LdS to projects. Added CSS to export the project designs. Use base64 encoding to create google docs documents in the background. Added google spreadsheet export format support. Added mini lds to projects preview. Added LdS title to project implementations. Fixed projects tools preview position. Add busy status to project implementation form. Fixed Factors and Concerns project tool subtype. Check for double relationship exception. Added LdS ability to write to a LdS according to project sharing.

+++ Ldshake/mod/lds/actions/lds/pdf_export.php
@@ -415,8 +415,79 @@ $contents = <<<EOD
         text-decoration: none;
     }
 
-    .draggable, .draggable-nonvalid { width: 90px; height: 90px; position: relative; float: left; margin: 10px 10px 10px 0; background-color: rgb(157,31,94); color: white !important; z-index: 10}
+/*** projects css ***/
+#droppable_grid {
+    width: 690px;
+    height: 616px;
+    padding: 0.5em;
+    border: 1px solid grey;
+    float: left;
+    background-color: white !important;
+    background-image: none;
+    position: relative;
+    z-index:0;
+}
 
+#ldproject_toolBar {
+    position: relative;
+    width: 250px;
+    height: 630px;
+    float:left;
+    border-style:
+    solid; border-color:
+    black; background-color: #F0F0F0 !important;
+    background-image: none;
+}
+
+.draggable, .draggable-nonvalid {
+    width: 75px;
+    height: 75px;
+    padding: 5px;
+    position: relative;
+    float: left;
+    margin: 8px 0px 0px 8px;
+    background-color: white;
+    color: white !important;
+    border: 1px solid #dddddd;
+    z-index: 10;
+}
+.ui-widget-content {background-image: none; }
+
+.ui-widget-header {background-image: none; background-color: green !important; border-color: black}
+
+.lds_att_popup {
+    top: 100px;
+    left: 50%;
+    z-index: 100000;
+    position: fixed;
+    background-color: #fff;
+    padding: 10px;
+    border: 10px solid rgba(120,120,120,0.42);
+    -moz-border-radius: 5px 5px 5px 5px;
+    -webkit-border-radius: 5px 5px 5px 5px;
+    border-radius: 5px 5px 5px 5px;
+    -webkit-background-clip: padding-box;
+    -moz-background-clip: padding;
+    background-clip: padding-box;
+    display: none;
+    font-size: 13px;
+}
+
+#lds_attachment_popup {
+    width: 500px;
+    top: 150px;
+    left: 50%;
+    margin-left: -260px;
+}
+
+.draggable > img[tooltype] {
+    width: inherit;
+}
+
+.draggable[tooltype_added="true"] {
+    padding: 0px 5px 10px 5px;
+    box-shadow: 2px 2px 0px #F0F0F0;
+}
 		</style>
 	</head>
 	<body>
@@ -426,25 +497,29 @@ $contents = <<<EOD
 </html>
 EOD;
 
-//Generate temp filenames
-$inname = "{$CONFIG->tmppath}HTM_{$doc_guid}.html";
-$outname = "{$CONFIG->tmppath}PDF_{$doc_guid}.pdf";
+if(isset($doc->file_pdf_guid)) {
+    $outname = ldshake_get_filepath($doc->file_pdf_guid);
+} else {
+    //Generate temp filenames
+    $inname = "{$CONFIG->tmppath}HTM_{$doc_guid}.html";
+    $outname = "{$CONFIG->tmppath}PDF_{$doc_guid}.pdf";
 
-if($doc->editorType == 'google_docs') {
-    $contents = $doc->description;
-    if($lds->license) {
-        $contents = str_replace('</body>', '<br />'.$license.'</body>', $contents);
-        $contents = str_replace('</style>', ' .license_banner{display: none;margin-top:20px;}'.'</style>', $contents);
+    if($doc->editorType == 'google_docs') {
+        $contents = $doc->description;
+        if($lds->license) {
+            $contents = str_replace('</body>', '<br />'.$license.'</body>', $contents);
+            $contents = str_replace('</style>', ' .license_banner{display: none;margin-top:20px;}'.'</style>', $contents);
+        }
     }
+
+    //Save the document data to the file
+    $handle = fopen($inname, "w");
+    fwrite($handle, $contents);
+    fclose($handle);
+
+    //Create the PDF (custom margins to respect web layout)
+    exec("{$CONFIG->pdf_converter_location} --dpi 102 -L 13mm -R 9mm $inname $outname");
 }
-
-//Save the document data to the file
-$handle = fopen($inname, "w");
-fwrite($handle, $contents);
-fclose($handle);
-
-//Create the PDF (custom margins to respect web layout)
-exec("{$CONFIG->pdf_converter_location} --dpi 102 -L 13mm -R 9mm $inname $outname");
 
 //Send the PDF to the client
 header("Content-disposition: attachment; filename=\"{$doc->title}.pdf\"");
@@ -452,5 +527,8 @@ header("Content-type: application/pdf");
 readfile($outname);
 
 //Remove the temp files
-unlink($inname);
-unlink($outname);
+if(isset($inname))
+    unlink($inname);
+
+if(!isset($doc->file_pdf_guid))
+    unlink($outname);
